# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ICICLE AI Vector Service",
    "description": "Vector storage and retrieval service for the ICICLE AI tenant. Clients provide\ntheir own pre-computed embeddings; the service handles storage, filtered search\nand reranking.\n\n**Collections** are broad domains (e.g. `biology`, `chemistry`), each backed by\nits own Qdrant collection and HNSW index. **Topics** are optional sub-categories\nwithin a collection (e.g. `human`, `plant`).\n\n**User isolation.** Every request is authenticated with a Tapis access token via\nthe `X-Tapis-Token` header, and `user_id` is taken from the token's\n`tapis/username` claim — never from the request body. Collections are physically\nshared, but every read, write, search and delete is filtered by `user_id`, so\nusers never see or affect each other's data.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/collections": {
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "List your collections",
        "operationId": "list_collections_v1_collections_get",
        "parameters": [
          {
            "name": "X-Tapis-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tapis-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "TapisToken": []
          }
        ]
      },
      "delete": {
        "tags": [
          "collections"
        ],
        "summary": "Delete ALL your embeddings across every collection",
        "description": "Irreversible: removes every point the caller owns, in every collection.\n\nOther users' data is untouched. Requires ``?confirm=true`` so that a stray\n``DELETE /v1/collections`` cannot wipe an account by accident.",
        "operationId": "purge_all_v1_collections_delete",
        "parameters": [
          {
            "name": "confirm",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Must be true. Guards against an accidental unqualified DELETE.",
              "default": false,
              "title": "Confirm"
            },
            "description": "Must be true. Guards against an accidental unqualified DELETE."
          },
          {
            "name": "X-Tapis-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tapis-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurgeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "TapisToken": []
          }
        ]
      }
    },
    "/v1/collections/{collection}": {
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "Get one collection's stats",
        "operationId": "read_collection_v1_collections__collection__get",
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection"
            }
          },
          {
            "name": "X-Tapis-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tapis-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "TapisToken": []
          }
        ]
      },
      "delete": {
        "tags": [
          "collections"
        ],
        "summary": "Delete your embeddings in one collection",
        "description": "Removes every point the caller owns in this collection.\n\nOther users keep theirs. The Qdrant collection is dropped only if it is left\nempty for everyone — reported as ``collection_dropped``.",
        "operationId": "delete_collection_v1_collections__collection__delete",
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection"
            }
          },
          {
            "name": "X-Tapis-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tapis-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "TapisToken": []
          }
        ]
      }
    },
    "/v1/collections/{collection}/embeddings": {
      "get": {
        "tags": [
          "collections"
        ],
        "summary": "List your embeddings in a collection",
        "description": "Cursor-paginated listing. Returns metadata only — never vectors.",
        "operationId": "list_embeddings_v1_collections__collection__embeddings_get",
        "parameters": [
          {
            "name": "collection",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Page size",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cursor from a previous page's next_offset",
              "title": "Offset"
            },
            "description": "Cursor from a previous page's next_offset"
          },
          {
            "name": "topic",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only embeddings with this topic",
              "title": "Topic"
            },
            "description": "Only embeddings with this topic"
          },
          {
            "name": "X-Tapis-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tapis-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "TapisToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkDeleteResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "collection": {
            "type": "string",
            "title": "Collection"
          },
          "deleted": {
            "type": "integer",
            "title": "Deleted"
          },
          "collection_dropped": {
            "type": "boolean",
            "title": "Collection Dropped",
            "description": "True when the underlying Qdrant collection was removed because no points from any user remained. Always false unless the deployment sets DROP_EMPTY_COLLECTIONS=true.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "collection",
          "deleted"
        ],
        "title": "BulkDeleteResponse"
      },
      "CollectionInfo": {
        "properties": {
          "collection": {
            "type": "string",
            "title": "Collection"
          },
          "points": {
            "type": "integer",
            "title": "Points"
          },
          "topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Topics"
          },
          "vector_dim": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector Dim"
          },
          "embedding_models": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Embedding Models"
          }
        },
        "type": "object",
        "required": [
          "collection",
          "points"
        ],
        "title": "CollectionInfo",
        "description": "A collection as seen by the requesting user.\n\n``points`` counts only the caller's own points, never every user's, so this\ncannot be used to infer how much data other users hold."
      },
      "CollectionList": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "collections": {
            "items": {
              "$ref": "#/components/schemas/CollectionInfo"
            },
            "type": "array",
            "title": "Collections"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "count",
          "collections"
        ],
        "title": "CollectionList"
      },
...
      "EmbeddingListResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "collection": {
            "type": "string",
            "title": "Collection"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset",
            "description": "Pass as ?offset= to fetch the next page. Null on the last page."
          },
          "embeddings": {
            "items": {
              "$ref": "#/components/schemas/EmbeddingRecord"
            },
            "type": "array",
            "title": "Embeddings"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "collection",
          "count",
          "embeddings"
        ],
        "title": "EmbeddingListResponse",
        "description": "A page of the caller's embeddings in one collection (metadata only)."
      },
      "EmbeddingRecord": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "collection": {
            "type": "string",
            "title": "Collection"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "vector_dim": {
            "type": "integer",
            "title": "Vector Dim"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "embedding_model": {
            "type": "string",
            "title": "Embedding Model"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "collection",
          "vector_dim",
          "created_at",
          "updated_at",
          "embedding_model"
        ],
        "title": "EmbeddingRecord"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PurgeResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "deleted": {
            "type": "integer",
            "title": "Deleted"
          },
          "collections_affected": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Collections Affected"
          },
          "collections_dropped": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Collections Dropped"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "deleted",
          "collections_affected",
          "collections_dropped"
        ],
        "title": "PurgeResponse",
        "description": "Result of deleting everything the caller owns, across all collections."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "TapisToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tapis-Token",
        "description": "ICICLE AI tenant Tapis access token. The username is read from the tapis/username claim and used to isolate your data."
      }
    }
  },
  "servers": [
    {
      "url": "https://icicleaivecserver.pods.icicleai.tapis.io",
      "description": "Production (ICICLE AI Tapis pod)"
    },
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    }
  ],
...
}